move env var settings up to top level module to prevent logging importing before env settings
